Extract App class
