+++ lib/main.dart
@@ -1,53 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:home_task/const.dart';
-import 'package:home_task/screens/home.dart';
+
+import 'app.dart';
 
 void main() {
-  runApp(const DemoApp());
-}
-
-class DemoApp extends StatelessWidget {
-  const DemoApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Home Task',
-      theme: ThemeData(
-        appBarTheme: const AppBarTheme(
-          centerTitle: false,
-          backgroundColor: primaryColor,
-          foregroundColor: Colors.black,
-          elevation: 2,
-          systemOverlayStyle: SystemUiOverlayStyle(
-            statusBarColor: Colors.transparent,
-          ),
-          iconTheme: IconThemeData(color: Colors.black),
-        ),
-        primaryColor: primaryColor,
-      ),
-      home: Scaffold(
-        appBar: AppBar(
-          leading: Container(),
-          titleSpacing: 10,
-          title: const Text(
-            titleMain,
-            style: TextStyle(
-              fontWeight: FontWeight.w600,
-              fontSize: 20,
-              shadows: <Shadow>[
-                Shadow(
-                  offset: Offset(0, 5),
-                  blurRadius: 9,
-                  color: Colors.black26,
-                ),
-              ],
-            ),
-          ),
-        ),
-        body: const Home(),
-      ),
-    );
-  }
+  runApp(const App());
 }
